Added NavigationBar.js
