+++ src/components/Events.js
@@ -14,6 +14,7 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import MuiAlert from '@mui/material/Alert';
+import NavigationBar from './NavigationBar';
 
 const { REACT_APP_API_ENDPOINT } = process.env;
 
@@ -186,6 +187,8 @@ const Events = () => {
 
   return (
     <div id="body">
+
+      <NavigationBar></NavigationBar>
 
       <Typography
         variant="h4"
