+++ src/components/UpdateProfile.js
@@ -2,6 +2,7 @@ import React, { useRef, useState } from "react"
 import { Form, Button, Card, Alert } from "react-bootstrap"
 import { useAuth } from "../contexts/AuthContext"
 import { Link, useHistory } from "react-router-dom"
+import NavigationBar from './NavigationBar';
 
 export default function UpdateProfile() {
   const emailRef = useRef()
@@ -43,6 +44,7 @@ export default function UpdateProfile() {
 
   return (
     <>
+    <NavigationBar></NavigationBar>
       <Card>
         <Card.Body>
           <h2 className="text-center mb-4">Update Profile</h2>
